requirements.txt
torch
torchvision
opencv-python>=4.8.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
pytesseract>=0.3.10
